updated to hapi 17.2.2 with new server,route option method, updated all npm packages, removed api/ from config to make swagger bit more clear for diffrentiating api's, moved all routes and components to api folder to imrove the overall code structure
