+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "typescript-node",
-  "version": "1.3.0",
+  "version": "1.3.1",
   "description": "",
   "license": "MIT",
   "repository": {
@@ -23,49 +23,46 @@
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
-    "boom": "~7.1.0",
+    "boom": "~7.2.0",
     "fs": "0.0.1-security",
     "good": "^8.1.0",
-    "good-console": "^7.0.1",
+    "good-console": "^7.1.0",
     "good-squeeze": "^5.0.2",
-    "hapi": "17.2.0",
+    "hapi": "17.2.2",
     "hapi-auth-jwt2": "^8.0.0",
     "hapi-swagger": "^9.1.1",
     "inert": "~5.1.0",
     "joi": "~13.1.2",
-    "jsonwebtoken": "^8.1.1",
-    "mongoose": "^5.0.4",
+    "jsonwebtoken": "^8.2.0",
+    "moment": "^2.21.0",
+    "moment-timezone": "^0.5.14",
+    "mongoose": "^5.0.9",
     "nconf": "^0.10.0",
-    "nodemon": "^1.11.0",
+    "nodemon": "^1.17.1",
     "path": "^0.12.7",
     "vision": "~5.3.1"
   },
   "devDependencies": {
-    "@types/bcryptjs": "^2.4.0",
+    "@types/bcryptjs": "^2.4.1",
     "@types/boom": "^7.1.1",
     "@types/chai": "^4.1.2",
-    "@types/hapi": "^17.0.0",
+    "@types/hapi": "^17.0.4",
     "@types/joi": "^13.0.5",
-    "@types/jsonwebtoken": "^7.2.0",
-    "@types/mocha": "^2.2.41",
-    "@types/mongoose": "^5.0.1",
+    "@types/jsonwebtoken": "^7.2.5",
+    "@types/mocha": "^2.2.48",
+    "@types/mongoose": "^5.0.6",
     "@types/nconf": "0.0.37",
-    "@types/node": "^9.4.5",
-    "@types/sinon": "^4.1.3",
+    "@types/sinon": "^4.3.0",
     "chai": "^4.1.2",
     "gulp": "^3.9.1",
     "gulp-env": "^0.4.0",
     "gulp-mocha": "^5.0.0",
-    "gulp-rimraf": "^0.2.1",
-    "gulp-shell": "^0.6.3",
-    "gulp-tslint": "^8.0.0",
-    "mocha": "^5.0.0",
-    "node": "9.5.0",
-    "sinon": "^4.3.0",
-    "tslint": "^5.4.3",
-    "typescript": "^2.4.1"
-  },
-  "engines": {
-    "node": "8"
+    "gulp-rimraf": "^0.2.2",
+    "gulp-shell": "^0.6.5",
+    "gulp-tslint": "^8.1.3",
+    "mocha": "^5.0.4",
+    "sinon": "^4.4.2",
+    "tslint": "^5.9.1",
+    "typescript": "^2.7.2"
   }
 }
