+++ src/configurations/config.test.json
@@ -6,7 +6,7 @@
     "port": 5000,
     "jwtSecret": "random-secret-password",
     "jwtExpiration": "1h",
-    "routePrefix": "/api",
+    "routePrefix": "",
     "plugins": [
       "logger",
       "jwt-auth"
