+++ src/server.ts
@@ -1,21 +1,20 @@
 import * as Hapi from 'hapi';
 import * as Boom from 'boom';
-import { IPlugin } from './plugins/interfaces';
-import { IServerConfigurations } from './configurations';
-import * as Tasks from './tasks';
-import * as Users from './users';
-import { IDatabase } from './database';
+import {IPlugin} from './plugins/interfaces';
+import {IServerConfigurations} from './configurations';
+import * as Tasks from './api/tasks';
+import * as Users from './api/users';
+import {IDatabase} from './database';
 
-export async function init(
-    configs: IServerConfigurations,
-    database: IDatabase
-): Promise<Hapi.Server> {
+export async function init(configs : IServerConfigurations, database : IDatabase) : Promise < Hapi.Server > {
     try {
         const port = process.env.PORT || configs.port;
         const server = new Hapi.Server({
             port: port,
             routes: {
-                cors: { origin: ['*'] }
+                cors: {
+                    origin: ['*']
+                }
             }
         });
 
@@ -24,24 +23,21 @@ export async function init(
         }
 
         //  Setup Hapi Plugins
-        const plugins: Array<string> = configs.plugins;
+        const plugins : Array < string > = configs.plugins;
         const pluginOptions = {
             database: database,
             serverConfigs: configs
         };
 
-        let pluginPromises: Promise<any>[] = [];
+        let pluginPromises : Promise < any > [] = [];
 
-        plugins.forEach((pluginName: string) => {
-            var plugin: IPlugin = require('./plugins/' + pluginName).default();
-            console.log(
-                `Register Plugin ${plugin.info().name} v${plugin.info().version}`
-            );
+        plugins.forEach((pluginName : string) => {
+            var plugin : IPlugin = require('./plugins/' + pluginName).default();
+            console.log(`Register Plugin ${plugin.info().name} v${plugin.info().version}`);
             pluginPromises.push(plugin.register(server, pluginOptions));
         });
 
         await Promise.all(pluginPromises);
-
 
         console.log('All plugins registered successfully.');
